Update the version of polygon.

+++ src/Module/Models/GymFilteringModel.cs
@@ -16,7 +16,7 @@ namespace SatelliteSite.ContestModule.Models
             var dict = new Dictionary<string, string>();
             if (Language != null) dict.Add("lang", Language);
             if (Problem != null) dict.Add("prob", Problem);
-            if (Verdict.HasValue) dict.Add("verd", Verdict.Value.ToString());
+            if (Verdict.HasValue) dict.Add("verd", ((int)Verdict.Value).ToString());
             if (dict.Count > 0) dict.Add("filter", "1");
             return dict;
         }
